Add fake mongoose emitter that can only fail + split stubs
New fake mongoose emitter for the purpose of testing the error handling
callback.

Required change was to split the stubs and emitters into seperate files
for best code reuse
